# swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Api Rest Adoptame",
    "description": "La documentacion",
    "version": "1.0.0"
  },
  "host": "restapi-adoptame.up.railway.app",
  "basePath": "/",
  "schemes": [
    "https"
  ],
  "paths": {
...
    "/api/v1.0/pets/": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "example": "any"
                },
                "typeId": {
                  "example": "any"
                },
                "breedId": {
                  "example": "any"
                },
                "typeHair": {
                  "example": "any"
                },
                "specialCares": {
                  "example": "any"
                },
                "castrated": {
                  "example": "any"
                },
                "gender": {
                  "example": "any"
                },
                "environment": {
                  "example": "any"
                },
                "tags": {
                  "example": "any"
                },
                "size": {
                  "example": "any"
                },
                "color": {
                  "example": "any"
                },
                "age": {
                  "example": "any"
                },
                "health": {
                  "example": "any"
                },
                "description": {
                  "example": "any"
                },
                "userId": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/v1.0/pets/{id}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "put": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "example": "any"
                },
                "typeId": {
                  "example": "any"
                },
                "breedId": {
                  "example": "any"
                },
                "typeHair": {
                  "example": "any"
                },
                "specialCares": {
                  "example": "any"
                },
                "castrated": {
                  "example": "any"
                },
                "gender": {
                  "example": "any"
                },
                "environment": {
                  "example": "any"
                },
                "tags": {
                  "example": "any"
                },
                "size": {
                  "example": "any"
                },
                "color": {
                  "example": "any"
                },
                "age": {
                  "example": "any"
                },
                "health": {
                  "example": "any"
                },
                "description": {
                  "example": "any"
                },
                "status": {
                  "example": "any"
                },
                "urlPhotosDb": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "delete": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
  }
}